implement Column scannable
Adds a new Scannable for Column structures. Columns have a pointer to a
ColumnDef struct, which can be manually constructed or fetched from the
Meta struct cache.

Columns also may have an alias, and the Size() and Scan() methods take
into account this alias when attempting to calculate buffer size and
copy() calls.