Implement Scannable for SELECT ... FROM
Adds a new `sqlb.Select()` function that accepts a list of Scannable
items and returns a `sqlb.Selectable` struct pointer. This
`sqlb.Selectable` itself implements Scannable, making it a reusable
component of the `sqlb` library.

`sqlb.Select()` analyzes the concrete types of the supplied list of
Scannable items and appends projected columns and subjects (elements of
the SQL FROM clause to itself). This means that we can call
`sqlb.Select()` with a Column, ColumnList, ColumnDef, Table, or
TableDef.

Closes Issue #6