Process interface{} arguments in Scan()
Renames Scannable interface to Element and modifies the Scan() method to
accept two slices and return two ints. The first slice is the buffer for
the SQL string that will be populated by the element. The second is the
pre-allocated array of variable arguments that will be supplied to the
eventual sql.Query(string, ...interface{}) call.

Adds a new method signature to the Scannable/Element interface called
ArgCount() which returns the number of arguments the element will copy
to the []interface{} parameter of Scan().