Prep for fixing table alias prefixing
In order to fix Issue #24, which requires prefixing column names in
output of a projected column or join condition with the alias of the
tbale or SelectClause that the column refers to, we needed to first
provide linkage between the *Column and the *Table, since *Table has
the alias, not the *TableDef.

In the process, we needed a way to uniquely identify Table and
TableDef structs. The solution was to make an id() method on each that
produced a unique identifier number by hashing either the Table.alias
or, if missing, the underlying TableDef's schema and name.