Add sqlb.Join() function and Selection interface
Adds a sqlb.Join() function that accepts two structs that implement the
new Selection interface and zero or more pointers to Expression structs.
This generic interface allows us to process both *Table and *TableDef
struct pointers as a component of a join operation and lays the
groundwork for enabled derived tables (SELECT in the FROM clause) which
will be *SelectClause struct pointers that will implement Selection.

Issue #23