Add support for SELECT ... FROM t1 JOIN t2 ON ...
Initial support for adding JOIN clauses to the SelectClause struct.
Currently does not properly handle >1 join arguments to Select()
properly but this is as good a place to stop as any because we're going
to need to fix the bug in Issue #28 before going much further in join
support.

Issue #23