Add Meta.NewTable() method
Adds a more refined way to create a TableDef that is associated to a
specific Meta struct. The Meta.NewTable() method takes a string table
name and returns a pointer to a TableDef that is associated with the
Meta.

Also adds the start of library reference documentation with the
beginning of information about constructing metadata.