Start reworking selectClause tests
The tests for the selectClause are not actually testing the selectClause
scan/size/argCount methods but rather are testing the Select()
function's behaviour. We want to pull the Select() function out into a
separate file (called query.go) with an exported Query struct being
returned. In order to do this, we need to make sure the select_test.go
tests are only testing the behaviour of the selectClause's
scan/size/argCount methods. This patch partially reworks those tests.