Working in a base set of metadata test fixtures
The goal is to move all of the unit tests (the tests that must be in
the sqlb package because they refer to internal structs) to using a
single set of metadata fixtures. These metadata fixtures are actually
currently defined in join_test.go and exist in the library's data
segment. I don't want any test-only data to be in the library's data
segment. I also think it's just better practice not to have any global
variables.