Move selectClause.String[Args]() to Query struct
Adds a new Query struct which will serve as the public interface to sqlb
query building library, not the private selectClause. Implements a
simple caching mechanism on the Query struct to not re-allocate the
query string and query arguments arrays unless the size and arg count
has changed on the underlying selectClause.