Ensure GROUP BY projections do not output alias
If a column is aliased, the resulting projection will output "tbl.col AS
alias" in the SELECT clause. However, in the GROUP BY clause, the
projection should be display just as "tbl.col".

In order to solve this issue, we add a new disableAliasScan() method to
the projection interface. This method sets the projection's alias to ""
and returns a reset functor that is used by the groupByClause's scan()
and size() methods to temporarily disable the outputting of the "AS
alias" extension into the SQL string.

Issue #40