Add simple test to select_test.go for single JOIN
Adds a test to select_test.go that demonstrates manually adding a single
joinClause to the selectClause's joins field. In doing so, this actually
uncovered an issue with Column.As() where that method was returning a
pointer to the Column itself, after mutating its own alias. That's not
what we want Column.As() to do. Instead, we want to return a *new*
Column struct with that new struct's alias set appropriately.