Prepare for derived table support
This branch is heading towards full support for derived tables -- i.e.
SELECT in the FROM clause. In order to get there, we need to abstract
the Query.sel field from a pointer to the concrete selectClause struct
into the selection interface.

We also want to support more than just the SELECT SQL statement. So, in
addition to the abstraction to the selection interface, we rename the
Query struct to SelectQuery and add a new public interface called Query
that will be used as the interface any struct that can output to a SQL
string.

Issue #10