Remove selectionId() from selection interface
After benchmarking some things, I believe the constant checking of
selectionId() values to compare selections to each other -- instead of
just checking to see if the selection's pointer address is the same as
another -- is a useless functionality that should not exist.

This functionality originally existed so that sqlb could know that if
the user did:

 t1 := meta.Table("users")
 t2 := meta.Table("users")
 t3 := meta.Table("users").As("u")

that sqlb would know t1 == t2 but t1 != t3.

Looking back, this is kinda silly. Users aren't going to use sqlb like
this.