Add safety measure guarding against nil ON expr
Ensures that trying to Join() or OuterJoin() a SelectQuery to a
selection and passing nil as the ON expression results in an error
instead of a segfault. Passing a nil ON expression is valid for CROSS
JOIN, but left a note to handle this situation in a later patch.

Issue #29