Add test case for compound ON clauses
The root cause of Issue #49 seems to be when the JOIN condition (ON
clause) contains a compound expression (e.g. and AND expression). In
this case, the SelectQuery.Join() method fails to determine that the
expression is a projection.

Add a test case for this and a commented-out correct SQL string.