Search joinClause structs for matching LEFT join
When looking for the selection that comprises the LEFT part of a
Join()'d selection, we were searching through the projections in the ON
expression and looking to see if any of the selections of the primary
selectClause struct matched the non-right part of the JOIN expression.
This worked fine for joins of the type A to B and A to C, for example
if we were doing something like this:

SELECT
  articles.id, users.name AS author, article_states.name AS state
FROM articles
JOIN users
  ON articles.author = users.id
JOIN article_states
  ON articles.state = article_states.id

However, if we attempted to do a join of the type A to B and B to C,
for example:

SELECT
  articles.id, users.name AS author, user_profiles.content AS author_profile
FROM articles
JOIN users
  ON articles.author = users.id
JOIN user_profiles
  ON users.id = user_profiles.user

Then we were failing to properly identify the users selection as the
LEFT part of the second join. This patch fixes that by searching
through the primary SelectQuery's selectClause's joins collection to
identify the appropriate LEFT side of the join.

Closes Issue #60