Fix SQL function projections in Select()
When doing the reference documentation for SQL function support, I
noticed that it wasn't possible to pass a *sqlFunc to the Select()
function. Well, it was possible to pass one, but the function would
transpose it into a literal value and an interpolation question mark.

This patch adds a *sqlFunc case to the type switch in the Select()
function, properly adding SQL functions as projections (and their
underlying selection) to the resulting SelectQuery's selectClause.

In the process of doing this, I realized that the `Count()` function
was producing a sqlFunc with no selection attribute so added a
selection parameter to that function and tests cases for it.

Closes Issue #64