Overhaul column/table structs and fix argCount()
Major overhaul of the way we represent table and column structs in sqlb. Gets
rid of the bifurcated Table/TableDef and Column/ColumnDef struct pairs and just
has a single Table and Column struct definition. When using Table.As() to
generate a new Table struct, we now copy all of the original Table struct's
columns and reference the new aliased Table struct in the cloned Column
structs.

This overhaul of the column/table metadata struct definitions enabled me to
debug what was going on in Issue #68. The derivedTable and Join structs were
not properly calculating argCount() by returning the argCount() of their
associated left, right, and from elements.

Fixes Issue #68
Closes Issue #69