add Insert() function
Adds a new Insert() function that returns an InsertQuery struct.
Similar to the SelectQuery struct, the InsertQuery struct implements
the Query interface.

The Insert() function accepts two arguments: a pointer to a Table
struct that will be the target of the INSERT SQL statement and a map,
keyed by string names of columns, of the values to use in the VALUES
clause of the INSERT statement.

Issue #67