Initial steps adding SQL dialect support
Changes the `sqlb.Meta` struct definition to contain a new dialect field
and updates the `sqlb.Reflect()` and `sqlb.NewMeta()` functions to
accept a Dialect enum/const value instead of a string representing the
DB driver name.

We'll hang dialect adaptation from this new field on the Meta struct,
allowing us to modify the output SQL buffer based on the dialect in the
meta.

Issue #56