Add support for dialect-specific interpolation
Major overhaul of the scan() interface in order to support
dialect-specific interpolation markers. Because PostgreSQL and other
RDBMS do not use the ? character for query parameter interpolation and
instead use a marker of $N, where N starts at 1, this means that the
[]byte buffer we use to inject the SQL string contents needs to expand
or contract slightly depending on the SQL dialect in use.

This patch changes the scan() interface from:

 scan([]byte, []interface{}) (int, int)

to:

 scan([]byte, []interface{}, *int) int

This change allowed me to pass a positional index for the argument
(which gets converted into a SQL dialect-specific interpolation marker
for each query parameter in the args slice) to a new
interpolationMarker() function that understands how to inject the
appropriate interpolation characters into the []byte buffer containing
the end SQL string.

Still need to add corresponding tests for PostgreSQL-specific string
output.

Issue #56