Clarify the TRIM() function
PostgreSQL's BTRIM() and MySQL's TRIM() function behave similarly when
supplied only a single argument: remove whitespace from the start and
end of the supplied argument.

However, the TRIM() function in both MySQL and PostgreSQL accepts a
location argument (TRAILING, LEADING and BOTH) and the PostgreSQL
variant of TRIM() requires a FROM in the function body. In preparation
for supporting these multi-argument variants of the TRIM() function,
this patch adds a TrimLocation type and adds a location and chars (for
the characters that should be trimmed) to the trimFunc struct.

Issue #56