Start of pretty-printing feature - add sqlScanner
The goal of this series is to rework the behaviour in sqlb where the
Dialect is passed around all over the place to one element after another
due to the way that the top-level Query interface-implementing structs
(like SelectQuery) construct the []byte buffer for printing SQL strings.

We'd like to be able to replace the current hard-coded SPACE character
as the delimiter for SQL clauses in the output SQL string from Query
structs. In order to do that, however, we need to clean up and decouple
the element interface from the thing that calls scan() on those
elements.

A new sqlScanner struct is introduced in this patch to help with that.
This first patch does little more than replace the dialect value in the
selectStatement and SelectQuery structs with a pointer to a sqlScanner
struct that we can use to hang formatting options and things like
Dialect (which affect the end string that is produced in the buffer) off
of.

Issue #51