Pass sqlScanner struct in scan() and size()
There were a number of structs that represent clauses or statements that
had a dialect field in them. This was done so that when calculating
buffer size or scanning the element into the output SQL []byte buffer,
the element could adapt itself accordingly. But now that we'd like to
adapt the output SQL string more, we want to instead pass a pointer to a
sqlScanner struct that will contain information for how the element
should interpret itself and scan itself into the buffer.

This meant changing the primary element scan() and size() interface
functions which of course meant changes to nearly all source files, but
the changes were pretty minimal and allowed me to completely remove the
setDialect() element interface function.

Issue #51