Remove Alias from api.Table and api.Column
Aliasing (the `[AS] {alias}` SQL clause) is not relevant for metadata
representation of a Table or Column. It is only relevant for
construction of the final SQL string as a field of a table_ref or
column_ref element.

This was leftover from before I separated the metadata representation of
Table and Column from the SQL grammar representation of table and column
identifiers.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>