add api.Projectable interface
To DRY things up in the `api/` package, this PR adds a new `Projectable`
interface that describes things that go in the SELECT clause of a SELECT
statement. In other words, things that can always produce a
DerivedColumn and can reference one or more Table or DerivedTable
(subquery) elements.

Updates the `api.AggregateFunction` struct first to implement the new
`api.Projectable` interface.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>