add SelectE() and other error-returning methods
Adds an error-returning `core/expr.SelectE()` variant function that does
not panic but instead returns a checkable error. For those users that
construct dynamic SQL statements using user-supplied input, we don't
want to panic but instead allow the calling code to check if the
constructed SQL expression is valid or not.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>